new flag to disable loop-idom
seaopt --enable-loop-idiom=[true,flag]

LoopIdiomRegognizer attempts to recognize and optimize common loops. In
particular, it optimizes initialization into a call to llvm.memset()
intrinsic.

This can be very useful for verification since memset can be handled in
more intelligent ways than arbitrary loops. However, we currently do not
handle it at all. So it is convenient to be able to disable the
optimization. At least for small examples.