Run LoopVectorizePass only if LoopVectorize flag is enabled.
The flag LoopVectorize is passed to the pass and it is used to tell
the vectorizer to vectorize the loop even if profiling recommends not
to do it. It is similar to the alwaysInline flag to the Inliner.

However, the LoopVectorize pass should NOT be run if flag
LoopVectorize is disabled. This is the help message from opt:

      -disable-loop-vectorization    - Disable the loop vectorization pass

(The option value is stored in LoopVectorize.)