fix(IPO): Annotation2Metadata produces valid metadata
In an upstream version of llvm, there is a new metadata ID for annotations. The id isn't define in llvm10, so
using it results in an 'Invalid ID' error. This commit fixes the bug by using a string identifier for the
metadata (string identifiers need not be built into llvm).